respaldo semanal 10 agosto 2015

+++ wp-content/plugins/enlighter/class/ObjectCache.php
@@ -27,10 +27,7 @@ class ObjectCache{
 	
 	// name of the cache file
 	private $_filename = null;
-	
-	// cache type (php and json are supported)
-	private $_cacheType = 'php';
-	
+
 	// cahce data availabe ?
 	private $_dataAvailable = false;
 	
@@ -56,10 +53,9 @@ class ObjectCache{
 	
 	/**
 	 * Load given file into internal storage
-	 * @param unknown $filename
 	 */
 	public function load(){
-		if (is_file($this->_filename) || is_readable($this->_filename)){
+		if (is_file($this->_filename) && is_readable($this->_filename)){
 			$this->_storage = require($this->_filename);
 			$this->_dataAvailable = true;
 		}
@@ -116,7 +112,7 @@ class ObjectCache{
 	
 	// clear cache
 	public function clear(){
-		if (is_file($this->_filename) || is_writable($this->_filename)){
+		if (is_file($this->_filename) && is_writable($this->_filename)){
 			unlink($this->_filename);
 		}
